Renamed and revised after feedback from @ben:
you can easily reduce the motivation to: having links point directly
network addresses makes it much more accessible and easy to share
content. The point you are currently making is a little vage and could
easily be done with DNS/NFS without "direct links".Options
Generally don't talk about what currently is the case or how it doesn't
work. that isn't the focus on the RFC, the focus is what the new feature
provides.Options
you can mention it in the motivation but, no need to reiterate that in
the requirements.Options
this shouldn't be an argument for the future but rather a description of
it.Options
the rest looks very good!Options
one side note: I learnt today that in order to retrieve a proper network
address we need to have a full DataIdentifier - So rather than saying
url by hash, let's call it "url via network address" and define the
"hash" as a base64 encoded serialisation of a DataIdentifier as defined
in
https://github.com/maidsafe/routing/blob/7c00dd14a2c4e4b2a3f7813a13edad119c0efa83/src/data/mod.rs#L111
maidsafe/routing
 - GitHub
routing - Routing - specialised storage DHT
Options
by doing it this way, this will also be automatically be possible with
any future additions of new data types (as they all always must be able
to be addressed as DataIdentifiers within the network). So if you could
update and use that instead, we are future proof in it.Options
I might sound harsh or direct, so I want to state this very clearly:
this is a great first draft! Thank you for doing this, it is very well
done.Options
I'd only like to cut down its length (as it is always easier to get
people to read shorter docs rather than longer ones ;) ) and focus it a
little clearer. But it is a great first go at it!Options
Does my feedback make sense? Let me know if you'd like me to clarify
anything or get more details or don't understand it or whatever helps
you move this forward.Options
One content question: so you don't want any mutable links at all?