Update 0061-safe-network-token-distribution.md
An update for this RFC as we round on some solutions. There are a lot of moving parts to consider; from technical capabilities, design propositions, operational considerations, to legal feedback, so thank you for your patience.

TOKEN EMISSION & GENESIS SUPPLY
---
The biggest change here is an answer to the question on whether we mint the total supply at Network Launch, and how and when the 70% of remaining tokens are created and then distributed.

Minting the total supply at launch necessitated either the Foundation or the Network itself holding and then distributing these tokens over time, both of which have significant challenges we want to avoid, chief among them security, but also the equitable distribution of a considerable slice of the economic power.

A reminder that the aims were A. gradual release of the Remaining Tokens, while at the same time B. having the tokens be in the custody of the Network.

The solution here is not to them minted up front at all, but rather have the Network emit them over time—creating them as the Network grows—and paying them out to Node operators as Resource Supply Rewards. 

While it still a chunky bit of work that is unlikely to be complete without delaying the Network inception, we are satisfied a reasonable design solution within reach, and one that can be implemented via a Network upgrade post launch.

So the Genesis supply, circulating at launch, will therefore be 30% of the Maximum Supply, and the remaining will only begin to come into being after a satisfactory, secure, and robust solution can be shown and rolled out; and then these tokens will drip out to Network contributors over an extended period, likely decades.

MINOR UPDATES
---
You'll also notice some other minor tweaks, such the language describing eligibility for Network Royalties as an App developer; more precision on the SNT allocation for shareholders; and directly stating that eMAID will be redeemable 1:1 for Safe Network Tokens, just the same as the original flavour MaidSafeCoin.

MAXIMUM SUPPLY
---
You'll note the change in the term from Total Supply, to Maximum Supply, which is in order to better articulate the changes around token emissions, and the proposed genesis supply.

One thing we're proposing remains from the RFC, is the change from the original Safecoin cap of 2^32 to the maximum supply of 4,525,524,120. This is to resolve the quirk from the original crowdsale that saw an overmint of MAID. 

But It's important to note that:

- No investors, crowdfund participants, nor MAID holders lose out.
- The long touted 1:1 Maid to SNT like-for-like swap remains; which is easy to describe and comprehend, and very useful to maintain for usability when on-ramping to the Network with MAID.
- Due to fundamental technical changes, tokens are no longer tied to network addresses. This allows for things like divisibility, but also gives us the freedom to specify a suitable maximum supply.
- The max supply of MAID has been known and advertised for nearly a decade now, displayed on every exchange and blockchain app going, along with the 1:1 redemption terms, so it’s not a surprise.
- It’s cleaner solution and we feel it’s easier to explain the the 5/10/15% allocations given to the initial distribution pools, and their share in the overall economy, than to repeatedly examine why MAID is now 10.37%.

We know this is a choice some of you won't like, but it's one between two aesthetic/ergonomic options, both of which arrive at the same result, so we've chosen what we feel is the path of least resistance.