Updating to reflect community feedback on initial token distribution
Updating the RFC to take on board concerns of the potential disadvantages to those holding MAID how purchased during the crowdsale prior to the over-issue tokens. This also means retaining the original proposed supply of 2^32 SNT. Woot!

Full breakdown of the alternatives consider at the end of the document.

Also other minot changes including procedure for uncontactable Shareholders.